A bit more refactoring
I propose we put utils-methods into `lib/common`, and distinguish them a bit more than just `utils`
