+++ lib/common/template-utils.js
@@ -1,0 +1,38 @@
+module.exports.processHandlebarFile = processHandlebarFile
+
+const Handlebars = require('handlebars')
+const debug = require('../debug').errors
+const { processFile } = require('./fs-utils')
+
+/**
+ * Reads a file, processes it (performing template substitution), and saves
+ * back the processed result.
+ *
+ * @param filePath {string}
+ * @param substitutions {Object}
+ *
+ * @return {Promise}
+ */
+async function processHandlebarFile (filePath, substitutions) {
+  return processFile(filePath, (rawSource) => processHandlebarTemplate(rawSource, substitutions))
+}
+
+/**
+ * Performs a Handlebars string template substitution, and returns the
+ * resulting string.
+ *
+ * @see https://www.npmjs.com/package/handlebars
+ *
+ * @param source {string} e.g. 'Hello, {{name}}'
+ *
+ * @return {string} Result, e.g. 'Hello, Alice'
+ */
+function processHandlebarTemplate (source, substitutions) {
+  try {
+    const template = Handlebars.compile(source)
+    return template(substitutions)
+  } catch (error) {
+    debug(`Error processing template: ${error}`)
+    return source
+  }
+}
