+++ lib/models/account-template.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const fs = require('fs-extra')
 const path = require('path')
 const mime = require('mime-types')
 const recursiveRead = require('recursive-readdir')
-const utils = require('../utils')
+const fsUtils = require('../common/fs-utils')
+const templateUtils = require('../common/template-utils')
 
 const RDF_MIME_TYPES = require('../ldp').RDF_MIME_TYPES
 const TEMPLATE_EXTENSIONS = [ '.acl', '.meta', '.json', '.hbs', '.handlebars' ]
@@ -62,13 +62,7 @@ class AccountTemplate {
    * @return {Promise}
    */
   static copyTemplateDir (templatePath, accountPath) {
-    return new Promise((resolve, reject) => {
-      fs.copy(templatePath, accountPath, (error) => {
-        if (error) { return reject(error) }
-
-        resolve()
-      })
-    })
+    return fsUtils.copyTemplateDir(templatePath, accountPath)
   }
 
   /**
@@ -130,7 +124,7 @@ class AccountTemplate {
    */
   processAccount (accountPath) {
     return this.readTemplateFiles(accountPath)
-      .then(files => Promise.all(files.map(path => utils.processHandlebarFile(path, this.substitutions))))
+      .then(files => Promise.all(files.map(path => templateUtils.processHandlebarFile(path, this.substitutions))))
   }
 
   /**
