+++ lib/server-config.js
@@ -6,7 +6,8 @@
 
 const fs = require('fs-extra')
 const path = require('path')
-const utils = require('./utils')
+const templateUtils = require('./common/template-utils')
+const fsUtils = require('./common/fs-utils')
 
 /**
  * Ensures that a directory has been copied / initialized. Used to ensure that
@@ -51,8 +52,8 @@ async function ensureWelcomePage (argv) {
   if (!fs.existsSync(existingIndexPage)) {
     console.log(argv)
     fs.mkdirp(serverRootDir)
-    await utils.copyTemplateDir(templates.server, serverRootDir)
-    await utils.processHandlebarFile(existingIndexPage, {
+    await fsUtils.copyTemplateDir(templates.server, serverRootDir)
+    await templateUtils.processHandlebarFile(existingIndexPage, {
       serverName: server ? server.name : host.hostname,
       serverDescription: server ? server.description : '',
       serverLogo: server ? server.logo : '',
