+++ lib/utils.js
@@ -13,16 +13,12 @@ module.exports.debrack = debrack
 module.exports.stripLineEndings = stripLineEndings
 module.exports.fullUrlForReq = fullUrlForReq
 module.exports.routeResolvedFile = routeResolvedFile
-module.exports.processHandlebarFile = processHandlebarFile
-module.exports.copyTemplateDir = copyTemplateDir
 
 const fs = require('fs-extra')
 const path = require('path')
 const $rdf = require('rdflib')
 const from = require('from2')
 const url = require('url')
-const Handlebars = require('handlebars')
-const debug = require('./debug').errors
 
 /**
  * Returns a fully qualified URL from an Express.js Request object.
@@ -149,64 +145,6 @@ function parse (data, baseUri, contentType, callback) {
   } catch (err) {
     return callback(err)
   }
-}
-
-/**
- * Reads a file, processes it (performing template substitution), and saves
- * back the processed result.
- *
- * @param filePath {string}
- * @param substitutions {Object}
- *
- * @return {Promise}
- */
-async function processHandlebarFile (filePath, substitutions) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(filePath, 'utf8', (error, rawSource) => {
-      if (error) {
-        return reject(error)
-      }
-
-      const output = processHandlebarTemplate(rawSource, substitutions)
-
-      fs.writeFile(filePath, output, (error) => {
-        if (error) {
-          return reject(error)
-        }
-        resolve()
-      })
-    })
-  })
-}
-
-/**
- * Performs a Handlebars string template substitution, and returns the
- * resulting string.
- *
- * @see https://www.npmjs.com/package/handlebars
- *
- * @param source {string} e.g. 'Hello, {{name}}'
- *
- * @return {string} Result, e.g. 'Hello, Alice'
- */
-function processHandlebarTemplate (source, substitutions) {
-  try {
-    const template = Handlebars.compile(source)
-    return template(substitutions)
-  } catch (error) {
-    debug(`Error processing template: ${error}`)
-    return source
-  }
-}
-
-async function copyTemplateDir (templatePath, targetPath) {
-  return new Promise((resolve, reject) => {
-    fs.copy(templatePath, targetPath, (error) => {
-      if (error) { return reject(error) }
-
-      resolve()
-    })
-  })
 }
 
 function serialize (graph, baseUri, contentType, callback) {
